requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.2.1
django-cors-headers==3.14.0
djangorestframework==3.14.0
joblib==1.2.0
numpy==1.24.3
pytz==2023.3
scikit-learn==1.2.2
scipy==1.10.1
sklearn==0.0.post4
sqlparse==0.4.4
threadpoolctl==3.1.0
